rename en.json to main.json
