+++ packages/tldraw/src/test/renderWithIntlProvider.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { IntlProvider } from 'react-intl'
-import messages_en from '~translations/en.json'
+import messages_en from '~translations/main.json'
 import messages_fr from '~translations/fr.json'
 
 export const renderWithIntlProvider = (children: React.ReactNode) => {
