+++ packages/tldraw/src/translations/main.json
